massive revamp of modules/function definitions location
