+++ magpie/management/__init__.py
@@ -1,8 +1,5 @@
-from service.resource import *
-from service.service import *
-
-
 def includeme(config):
     config.include('management.group')
     config.include('management.user')
     config.include('management.service')
+    config.include('management.resource')
