+++ magpie/management/group/group_utils.py
@@ -1,5 +1,7 @@
 from api_requests import *
 from api_except import *
+from models import resource_tree_service
+from ziggurat_definitions import *
 
 
 def format_group(group):
