+++ magpie/models.py
@@ -1,24 +1,8 @@
 from pyramid.security import Allow, Everyone
 from pyramid.security import ALL_PERMISSIONS
-from pyramid.httpexceptions import HTTPInternalServerError
-from ziggurat_foundations.models.external_identity import ExternalIdentityMixin
-from ziggurat_foundations.models.group import GroupMixin
-from ziggurat_foundations.models.group_permission import GroupPermissionMixin
-from ziggurat_foundations.models.group_resource_permission import GroupResourcePermissionMixin
-from ziggurat_foundations.models.resource import ResourceMixin
-from ziggurat_foundations.models.user import UserMixin
-from ziggurat_foundations.models.user_group import UserGroupMixin
-from ziggurat_foundations.models.user_permission import UserPermissionMixin
-from ziggurat_foundations.models.user_resource_permission import UserResourcePermissionMixin
-from ziggurat_foundations import ziggurat_model_init
-from sqlalchemy.ext.declarative import declarative_base
-from ziggurat_foundations.models.base import get_db_session
+from sqlalchemy.ext.declarative import declarative_base, declared_attr
 import sqlalchemy as sa
-from ziggurat_foundations.permissions import permission_to_pyramid_acls
-from sqlalchemy.ext.declarative import declared_attr
-from ziggurat_foundations.models import groupfinder as gf
-from ziggurat_foundations.models.services.resource_tree import ResourceTreeService
-from ziggurat_foundations.models.services.resource_tree_postgres import ResourceTreeServicePostgreSQL
+from ziggurat_definitions import *
 from api_except import *
