+++ magpie/ziggurat_definitions.py
@@ -1,0 +1,22 @@
+# import required definitions
+from ziggurat_foundations import ziggurat_model_init
+from ziggurat_foundations.models import groupfinder as gf
+from ziggurat_foundations.models.base import get_db_session
+from ziggurat_foundations.models.external_identity import ExternalIdentityMixin
+from ziggurat_foundations.models.group import GroupMixin
+from ziggurat_foundations.models.group_permission import GroupPermissionMixin
+from ziggurat_foundations.models.group_resource_permission import GroupResourcePermissionMixin
+from ziggurat_foundations.models.resource import ResourceMixin
+from ziggurat_foundations.models.services.external_identity import ExternalIdentityService
+from ziggurat_foundations.models.services.group import GroupService
+from ziggurat_foundations.models.services.group_resource_permission import GroupResourcePermissionService
+from ziggurat_foundations.models.services.resource import ResourceService
+from ziggurat_foundations.models.services.resource_tree import ResourceTreeService
+from ziggurat_foundations.models.services.resource_tree_postgres import ResourceTreeServicePostgreSQL
+from ziggurat_foundations.models.services.user import UserService
+from ziggurat_foundations.models.services.user_resource_permission import UserResourcePermissionService
+from ziggurat_foundations.models.user import UserMixin
+from ziggurat_foundations.models.user_group import UserGroupMixin
+from ziggurat_foundations.models.user_permission import UserPermissionMixin
+from ziggurat_foundations.models.user_resource_permission import UserResourcePermissionMixin
+from ziggurat_foundations.permissions import permission_to_pyramid_acls
